fix: request token list if necessary
