+++ src/stores/bridgeStore.ts
@@ -1,6 +1,9 @@
 import { create } from "zustand"
 
-import { NETWORKS } from "@/constants"
+import { tokenListUrl } from "@/apis/dynamic"
+import { NATIVE_TOKEN_LIST, NETWORKS } from "@/constants"
+import { USER_TOKEN_LIST } from "@/constants/storageKey"
+import { loadState } from "@/utils/localStorage"
 
 type TransactionType = "Deposit" | "Withdraw"
 
@@ -26,6 +29,7 @@ interface BridgeStore {
   withDrawStep: WithDrawStep
   txResult: TxResult
   isNetworkCorrect: boolean
+  tokenList: Array<Token>
 
   changeFromNetwork: (network: Network) => void
   changeToNetwork: (network: Network) => void
@@ -33,6 +37,7 @@ interface BridgeStore {
   changeTxResult: (txResult: TxResult | null) => void
   changeWithdrawStep: (withDrawStep: WithDrawStep) => void
   changeIsNetworkCorrect: (isNetworkCorrect: boolean) => void
+  fetchTokenList: () => Promise<void>
 }
 
 const useBridgeStore = create<BridgeStore>()((set, get) => ({
@@ -43,6 +48,37 @@ const useBridgeStore = create<BridgeStore>()((set, get) => ({
   withDrawStep: "1",
   txResult: null,
   isNetworkCorrect: true,
+  tokenList: [],
+
+  fetchTokenList: async () => {
+    const { tokens: tokensListTokens } = await scrollRequest(tokenListUrl)
+    const currentUserTokens = loadState(USER_TOKEN_LIST) || []
+
+    const combinedList = [...NATIVE_TOKEN_LIST, ...tokensListTokens, ...currentUserTokens]
+    const uniqueList = combinedList.reduce(
+      (accumulator, token) => {
+        // If the token doesn't have an address, consider it a native token and add it directly
+        if (!token.address) {
+          accumulator.result.push(token)
+          return accumulator
+        }
+        // Convert address to lowercase for case-insensitive deduplication
+        const lowercaseAddress = token.address.toLowerCase()
+        // Create a key combining address and chainId to ensure different chainIds are not deduplicated
+        const key = `${lowercaseAddress}-${token.chainId}`
+
+        if (!accumulator.seen[key]) {
+          accumulator.seen[key] = true
+          accumulator.result.push(token)
+        }
+        return accumulator
+      },
+      { seen: {}, result: [] },
+    ).result
+    set({
+      tokenList: uniqueList,
+    })
+  },
 
   changeHistoryVisible: value => {
     set({
